更新

+++ app/src/main/java/com/miittech/you/net/response/UserInfoResponse.java
@@ -282,24 +282,24 @@ public class UserInfoResponse extends BaseResponse{
              * lasttime :
              */
 
-            private int lat;
-            private int lng;
+            private double lat;
+            private double lng;
             private String addr;
             private String lasttime;
 
-            public int getLat() {
+            public double getLat() {
                 return lat;
             }
 
-            public void setLat(int lat) {
+            public void setLat(double lat) {
                 this.lat = lat;
             }
 
-            public int getLng() {
+            public double getLng() {
                 return lng;
             }
 
-            public void setLng(int lng) {
+            public void setLng(double lng) {
                 this.lng = lng;
             }
